Fixes for github actions lik unit tests (#992)
* use int

* Update esri test data source

* Run tests on DEBUG loglevel

* Use np.int32

+++ pygeoapi/provider/xarray_.py
@@ -519,7 +519,7 @@ class XarrayProvider(BaseProvider):
             dt = np.array([time_diff.values.astype('timedelta64[{}]'.format(x))
                            for x in ['Y', 'M', 'D', 'h', 'm', 's', 'ms']])
 
-            return str(dt[np.array([x.astype(np.int) for x in dt]) > 0][0])
+            return str(dt[np.array([x.astype(np.int32) for x in dt]) > 0][0])
         else:
             return None
